requirements.txt
pillow
pickleshare
numpy
Flask
urllib3
gunicorn
